requirements.txt
Flask==2.0.3
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
python-barcode==0.14.0
python-dotenv==0.20.0
SQLAlchemy==1.4.46
Werkzeug==2.0.3
WTForms==3.0.0
json_handler==2.0.1
numpy
torch
matplotlib
alpha-vantage==1.2.1
polygon-api-client
textblob
setuptools
requests
